Add FARM framework
Building upon the transformers repo, FARM is a home for all species of pretrained language models (e.g. BERT) that can be adapted to different domain languages or down-stream tasks. Standardized interfaces for language models and prediction heads allow easy application for practitioners and fast iteration of experiments. Additional experiment tracking and visualizations support you along the way to have a fast proof-of-concept.